pyproject.toml
[tool.poetry]
name = "configcraft-ai"
version = "0.1.0"
description = ""
authors = ["Fanyi Zhao <fzhao12@stevens.edu>"]
license = "MIT"
readme = "README.md"
packages = [{include = "configcraft"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
#mistralai = "^0.1.8"
llama-index = "^0.10.30"
python-dotenv = "^1.0.1"
pymupdf = "^1.24.1"
# llama-index-embeddings-mistralai = "^0.1.4"
# llama-index-llms-mistralai = "^0.1.10"
streamlit = "^1.33.0"
watchdog = "^4.0.0"
llama-index-agent-openai = "^0.2.2"
llama-index-readers-file = "^0.1.19"
llama-index-vector-stores-postgres = "^0.1.7"
llama-index-embeddings-huggingface = "^0.2.0"
llama-index-llms-llama-cpp = "^0.1.3"
jupyter = "^1.0.0"
llama-cpp-python = "^0.2.74"
psycopg2-binary = "^2.9.9"
pgvector = "^0.2.5"
asyncpg = "^0.29.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.30"}
greenlet = "^3.0.3"
llama-index-llms-ollama = "^0.1.3"
llama-index-embeddings-ollama = "^0.1.2"
llama-index-readers-web = "^0.1.13"
llama-index-readers-github = "^0.1.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
